Add structopt docs and config
This adds the initial scaffolding for processing some of the
command line arguments. As a first pass, I just want to be
able to get simple pattern substitutions working on a single
file, so the CLI arguments reflect that. Otherwise nothing
function really happens yet.

I added a snapshot test for the output of `gsub --help`.
This doesn't really test anything, it's just a way to
document (for myself) what the current CLI options look like.