Just like, a bunch of refactoring
The original intention of this commit was to add the ability to
filter out certain files.

There's some of that, but it's a NOOP as far as main() is concerned.

Most of this is just, like, a BUNCH of refactoring (especially
on file_iterator.rs) to be more idiomatic with iterators and such.